переделка 3
