--- src/components/CardFilm copy/CardFilm.tsx
@@ -1,12 +1,0 @@
-
-
-function CardFilm() {
-	return (
-					<div>
-							alt="icon rating"
-					</div>
-					
-	);
-}
-
-export default CardFilm;
