+++ src/components/CardFilm/CardFilm.module.css
@@ -8,16 +8,16 @@
 
 	backdrop-filter: blur(40px);
 	display: inline-flex;
-	padding: 8px 8px 16px 8px;
+	padding: 8px 8px 8px 8px;
 	flex-direction: column;
 	align-items: flex-start;
-	gap: 16px;
+	gap: 20px;
 	color: var(--Grey-50, #ebeef5);
 	font-weight: 600;
 	line-height: 24px;
 	letter-spacing: 0.32px;
 	font-family: 'Poppins', sans-serif;
-	font-size: 16px;
+	font-size: 20px;
 	font-style: normal;
 	width: 266px;
 	height: 504px;
@@ -26,6 +26,13 @@
 .card-button:hover {
 	background: rgba(20, 22, 27, 0.8);
 	backdrop-filter: blur(40px);
+}
+
+.rating {
+	position: absolute;
+	top: 16px;
+	left: 16px;
+	z-index: 2;
 }
 
 .poster {
@@ -38,32 +45,4 @@
 	width: 100%;
 	height: 100%;
 	object-fit: cover;
-}
-
-.favorite-button {
-	border: none;
-	cursor: pointer;
-	background: none;
-	display: flex;
-	align-items: center;
-	gap: 8px;
-	color: var(--Primary-400, #7b6ef6);
-	font-family: Poppins;
-	font-size: 16px;
-	font-style: normal;
-	font-weight: 600;
-	line-height: 24px;
-	letter-spacing: 0.32px;
-}
-
-.favorite-button:hover {
-	color: #514c7c;
-}
-
-.favorite-button_active {
-	color: var(--Success-400, #37d8a7);
-}
-
-.favorite-button_active:hover {
-	color: #365c50;
 }
