+++ src/components/FavoriteButton/FavoriteButton.tsx
@@ -1,34 +1,34 @@
 import styles from './FavoriteButton.module.css';
-import { useState, MouseEvent } from 'react';
+import { MouseEvent } from 'react';
 import cn from 'classnames';
 
+import { useDispatch, useSelector } from 'react-redux';
+import { AppDispath, RootState } from '../../store/store';
+import { cartActions } from '../../store/favoriteList.slice';
+
 interface FavoriteButtonProps {
-	onClick?: () => void;
+	id: number;
 }
 
-function FavoriteButton({ onClick }: FavoriteButtonProps) {
-	const [isFavorite, setIsFavorite] = useState(false);
+function FavoriteButton(props: FavoriteButtonProps) {
+	const dispatch = useDispatch<AppDispath>();
+	const cartItems = useSelector((state: RootState) => state.cart.items);
+	const isFavorite = cartItems.some(item => item.id === props.id);
 
-	const handleButtonClick = (e: MouseEvent<HTMLDivElement>) => {
+	const toggleFavorite = (e: MouseEvent) => {
 		e.stopPropagation();
-		setIsFavorite(!isFavorite);
-		if (onClick) {
-			onClick();
+		e.preventDefault();
+		if (isFavorite) {
+			dispatch(cartActions.delete(props.id));
+		} else {
+			dispatch(cartActions.add(props.id));
 		}
 	};
 
 	return (
-		<div
-			className={cn(styles['favorite-button'], {
-				[styles['favorite-button_active']]: isFavorite
-			})}
-			onClick={handleButtonClick}
-		>
-			<div>
-				<img
-					src={isFavorite ? '/like active.svg' : '/like.svg'}
-					alt="icon rating"
-				/>
+		<div className={cn(styles['favorite-button'], {[styles['favorite-button_active']]: isFavorite})} onClick={toggleFavorite} >
+			<div className={styles['icon']}>
+				<img src={isFavorite ? '/like active.svg' : '/like.svg'} alt="icon rating" />
 			</div>
 			<div>{isFavorite ? 'В избранном' : 'В избранное'}</div>
 		</div>
