+++ src/components/Rating/Rating.module.css
@@ -1,9 +1,4 @@
-.rating-button {
-	position: absolute;
-	top: 16px;
-	left: 16px;
-	z-index: 2;
-
+.rating {
 	border: none;
 	cursor: pointer;
 	display: inline-flex;
@@ -20,11 +15,8 @@
 	font-weight: 400;
 	line-height: 24px;
 }
-.rating-button:hover {
-	background: rgba(109, 101, 101, 0.65);
-}
 
-.rating-button__icon {
+.rating__icon {
 	display: flex;
 	width: 16px;
 	height: 16px;
