--- src/interfaces/auth.interface.ts
@@ -1,3 +1,0 @@
-export interface LoginResponse {
-	access_token: string;
-}
