+++ src/pages/PageMovie/PageMovie.module.css
@@ -11,9 +11,30 @@
 	color: #ebeef5;
 }
 
+.description {
+	display: flex;
+	gap: 24px;
+}
+
 .poster {
-	border-radius: 24px;
 	width: 480px;
 	height: 720px;
-	background-size: cover;
+}
+
+.poster img {
+	width: 100%;
+	height: 100%;
+	object-fit: cover;
+	border-radius: 24px;
+}
+
+.description__text {
+	display: flex;
+	flex-direction: column;
+	gap: 24px;
+}
+
+.description__variable {
+	display: flex;
+	gap: 24px;
 }
